app/requirements.txt
flask~=2.3.2
flask[async]
requests~=2.0
pymongo~=4.3.3
python-dotenv
pandas